feat!: ESLint 8 and jsonc-eslint-parser (#40)
Fixes #39; fixes #43.

Moves to the external `jsonc-eslint-parser` because:

* #43: it's a good ecosystem standard
* ESLint's RuleTester doesn't support this usage of processors
(https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/pull/40#issuecomment-1726969178)

Also removes the `disparity` dependency, in favor of directly telling
users to run the auto-fixer. This is the strategy other plugins such as
[`eslint-plugin-simple-import-sort`](https://github.com/lydell/eslint-plugin-simple-import-sort)
take. It's simpler to implement, makes for easier-to-read error
messages, and reduces the size of `node_modules`.