refactor: resolve multiple eslint plugin violations
This change resolves all of the violations revealed by #708, except for one.  The messages used to report violations in `valid-package-def` are coming from PJV and not maintained by the plugin.  So unless there's an attempt to map PJV errors to statically defined messages within the rule, that one will need to continue using `message`