feat(valid-package-definition): rename rule from valid-package-def (#711)
<!-- 👋 Hi, thanks for sending a PR to eslint-plugin-package-json! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

-   [x] Addresses an existing open issue: fixes #132 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change deprecates the `valid-package-def` rule in favor of a new
`valid-package-definition` rule, providing the same behavior. The
deprecated version of the rule will be removed in the next major
release.

I assumed you wanted to roll this change out in a non-breaking way
initially? If that's not the case and you want to major, I can remove
the deprecated rule. If you do prefer this approach, then it may be
worth documenting the follow-up so that when it's time to major, the
deprecated rule is dropped?

Closes #132