ci: adjust release page generation
This change is a POC to address https://github.com/JoshuaKGoldberg/create-typescript-app/issues/1913.  Since it's not something we can really test locally, I figured we test it out here, and then, if successful, I can make the updates upstream.

Here we've changed the release-it config from json to js, which allows us to create a custom function for `github.releaseNotes`.  Using the context provided by release-it, we can filter by groups that we want to keep, and organize them into the same groups that we're organizing for the CHANGELOG (via conventional-changelog).  I updated the prepare action to use node 22, which will allow us to use the `Object.groupBy` api (introduced in v21).  Since 22 is now LTS, this felt like a reasonable change.