ci: fix the release it config to only show user-facing changes
This change adjusts the release it config so that the output generated from the changelog plugin is used for the github release page.

I discussed this here https://github.com/release-it/release-it/issues/1025#issuecomment-2617948522, and @webpro theorized it was the autoGenerate that overrode the changelog output.

I created a test repo to verify and confirmed that that, did indeed fix the issue: https://github.com/michaelfaith/mf-cta-testing/releases
2.0.2 is after this change, which included both a chore change and fix change, but changelog and GH release page only showed the bug fix (notice also the group heading)