chore: add build pre-step to doc update rule
This change adds a build pre-step to the eslint doc update script.  There's never really a case where you'd not want to have a build happen when running the doc update command, and when running locally, it's not super obvious that it's necessary to run a build first.  Baking it into the update script removes that ambiguity.