feat!: refactor plugin export to conform to latest conventions
This change refactors the plugin's exports to conform to the latest standards recommended by the ESLint team.

The plugin is still the default export, but it now has configs attached to it as well, but the legacy rc-based config (now called `'legacy-recommended'`) and the `recommended` flat-config.

For users of the flat config, you'll need to make the following changes.

```js
- import packageJson from "eslint-plugin-package-json/configs/recommended";
+ import packageJson from "eslint-plugin-package-json";

export default [
-	 packageJson,
+	 packageJson.configs.recommended,
];
```

For users of the legacy config, you'll need to make the following changes.

```js
module.exports = {
	overrides: [
		{
- 			extends: ["plugin:package-json/recommended"],
+ 			extends: ["plugin:package-json/legacy-recommended"],
			files: ["package.json"],
			parser: "jsonc-eslint-parser",
		},
	],
};
```

BREAKING CHANGE