feat!: refactor plugin export to conform to latest conventions (#957)
## PR Checklist

-   [x] Addresses an existing open issue: fixes #752 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change refactors the plugin's exports to conform to the latest
standards recommended by the ESLint team.

The plugin is still the default export, but it now has configs attached
to it as well, but the legacy rc-based config (now called
`'legacy-recommended'`) and the `recommended` flat-config.

For users of the flat config, you'll need to make the following changes.

```diff
- import packageJson from "eslint-plugin-package-json/configs/recommended";
+ import packageJson from "eslint-plugin-package-json";

export default [
-   packageJson,
+   packageJson.configs.recommended,
];
```

For users of the legacy config, you'll need to make the following
changes.

```diff
module.exports = {
  overrides: [
    {
-     extends: ["plugin:package-json/recommended"],
+     extends: ["plugin:package-json/legacy-recommended"],
      files: ["package.json"],
      parser: "jsonc-eslint-parser",
    },
  ],
};
```

BREAKING CHANGE

Closes #752