feat(valid-package-definition): add ignoreProperties option (#1077)
## PR Checklist

- [x] Addresses an existing open issue: fixes #1076 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `ignoreProperties` option to the
`valid-package-definition` rule, allowing you to provide a list of
top-level properties to suppress violation reports for. This can be
useful when using this rule alongside any of the more granular `valid-*`
rules, so that it's not double reporting violations.