feat(valid-bin): add new rule for validating bin
This change adds a new `valid-bin` rule that uses `validateBin` from `package-json-validator` to identify errors.

This will need to wait to merge after #1077, so that we're not double-reporting validation errors on `bin` with `valid-package-definition`.

Lastly, this is the first new rule that employs our new approach of using the more granular validation functions from `package-json-validator` and can serve as a model for the rest of the `valid-*` rules.