feat(valid-bin): add new rule for validating bin (#1082)
## PR Checklist

- [x] Addresses an existing open issue: fixes #818 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `valid-bin` rule that uses `validateBin` from
`package-json-validator` to identify errors.

This will need to wait to merge after #1077, so that we're not
double-reporting validation errors on `bin` with
`valid-package-definition`.

Lastly, this is the first new rule that employs our new approach of
using the more granular validation functions from
`package-json-validator` and can serve as a model for the rest of the
`valid-*` rules.