fix(valid-local-dependency): don't flag archives as invalid
This change fixes an error with the `valid-local-dependency` rule, where it would mis-report local paths to archives as invalid.  It now correctly treats valid paths to an archive as valid.  It does _not_ attempt to unpack the archive or validate its contents in any way.  Just that it's located where the package.json says it is.  An archive with invalid contents would manifest errors in other ways.