refactor: add utility function for generating basic `valid-` rules
This change adds a new utility function in the same vein as the one we use to generate simple `require-` rules, but for simple `valid-` rules.  If the valid- rule doesn't need any options, and is simply using `package-json-validator` to do the validation, then it can be generated using this new function.