feat(no-empty-fields): Add `ignoreProperties` option
Various tools store configuration in package.json, and some may make use
of non-empty arrays or objects as part of that configuration.

It's not reasonable for eslint-plugin-package-json to know about all
such tools, but it is reasonable to give users the ability to configure
the rule to ignore properties for tools that they make use of.