feat: add new `require-` rules for `bugs`, `bundleDependencies`, `dependencies`, and more
This change adds new `require-` rules for the following properties.

- `bugs`
- `bundleDepdencies`
- `dependencies`
- `devDependencies`
- `optionalDependencies`
- `peerDependencies`

I also refactored the tests for these, so that we don't have to add a new test file for every new `require-` prop we add.  `require-properties` will just pick it up automatically.  Now, it's as simple as adding a new entry to the `properties` array, and adding a new documentation file, to introduce a new `require-` rule.