fix(sort-collections): align sorting of scripts with `prettier-plugin-packagejson` (#1203)
<!-- 👋 Hi, thanks for sending a PR to eslint-plugin-package-json! 🗂
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #753 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change updates how we're sorting script properties in
`sort-collections`, to align with how `prettier-plugin-packagejson`
sorts it. The main gap in what we had in our bespoke implementation and
what we have now, is that pre/post lifecycle scripts were not being
grouped together if the non-prefixed version of the script was not also
included. For example, it's rare to have an `install` script in a
package.json, but super common to have a `postinstall`. Now `preinstall`
and `postinstall` are sorted correctly in the place where `install`
would normally be sorted, with `preinstall` coming first and
`postinstall` coming second.

In order to achieve this, I replaced the bespoke script sorting
implementation with a call to `sort-package-json`, which is what the
prettier plugin also uses to sort.