build: change relative import extensions to ts
This change moves all of our relative path imports to use `.ts` extensions instead of `.js`.  In order to do this, and avoid a bug in `tsup`, we're moving to use `tsdown` for bundling.  There should be no user-facing changes from this update.

This simply sets us up to use node's native type stripping for our own linting within the repo.

I also changed the ci workflow to default node version to the latest lts version.  This doesn't affect the matrix used for the test job.