refactor: infer rule options type from json schema (#1258)
<!-- 👋 Hi, thanks for sending a PR to eslint-plugin-package-json! 🗂
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #1259
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/eslint-plugin-package-json/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

I'm not sure if this PR requires an issue. I'm happy to open one if it
is 🙂

I've noticed that rule options types are written manually and thus must
be in sync with JSON schema. This is a QoL PR that makes type inference
automatic using
[`json-schema-to-ts`](https://www.npmjs.com/package/json-schema-to-ts).
This change also allowed `meta.defaultOptions` to be typed.

If for some reason `json-schema-to-ts` is not able to correctly infer
the type, there's an escape hatch that allows manually passing the
options type to the first type parameter of `createRule`.