feat(valid-author): create more precise reports
This change incorporates the new rich output from `package-json-validator`'s `validateAuthor` function, which allows us to report at a more granular level for complex objects.  We're now reporting on the individual properties of Author objects, when violations occur for those properties.

I'm also treating this rule as a prototype for the next generate of factory function for `valid-*` rules, eventually replacing what's in `createSimpleValidPropertyRule`.  In a follow-up change, I'll introduce a new version of that, for the rules powered by the new `package-json-validator` `Result` type.