fix(repository-shorthand): support repo provider prefix
This change updates all fixers to support different repo providers, and the providers prefix in general.  Now if you have repo defined in object syntax with either bitbucket, gitlab, github, or gist urls, the fixer will construct a shorthand string using the corresponding provider prefix.  On the other side, if you have repo defined as a shorthand and have configured to prefer object, it'll use the correct url based on the provider prefix, if specified (assume github if not).